+++ js/main.js
@@ -2,22 +2,7 @@ import {makeOffers} from './data.js';
 import {getRandomArrayElement} from './utils.js';
 import {getRandomPositiveInteger} from './utils.js';
 import {getRandomPositiveFloat} from './utils.js';
-
-
-const getRandomFeatures = (count, array) => {
-  const newFeatures = [];
-  if (count >= array.length) {
-    return array;
-  }
-  for (let i = 0; i < count; i++) {
-    let newElement = array[Math.floor(Math.random() * array.length)];
-    while (newFeatures.includes(newElement)) {
-      newElement = array[Math.floor(Math.random() * array.length)];
-    }
-    newFeatures.push(newElement);
-  }
-  return newFeatures;
-};
+import {getRandomFeatures} from './utils.js';
 
 
 const makeGenerator = (count) => {
@@ -27,6 +12,4 @@ const makeGenerator = (count) => {
 };
 
 const getAvatar = makeGenerator(10);
-
-export {getRandomFeatures, makeGenerator, getAvatar};
